+++ NEWS.md
@@ -1,3 +1,8 @@
+# tzara 0.0.9
+
+* Added arguments `"seq_id"` and `"filename"` to `dadamap()`, to make it more
+  user-friendly.  Use of `add_dada_names()` is now unnecessary.
+
 # tzara 0.0.8
 
 * Added a `NEWS.md` file to track changes to the package.